+++ src/opera_rtc_s1_browse/__init__.py
@@ -1,9 +1,0 @@
-"""A tool for create OPERA RTC S1 browse images"""
-
-from importlib.metadata import version
-
-__version__ = version(__name__)
-
-__all__ = [
-    '__version__',
-]