+++ tests/test_entrypoints.py
@@ -1,3 +1,0 @@
-def test_opera_rtc_s1_browse(script_runner):
-    ret = script_runner.run('python', '-m', 'opera_rtc_s1_browse', '-h')
-    assert ret.success