+++ CHANGELOG.md
@@ -6,5 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [PEP 440](https://www.python.org/dev/peps/pep-0440/)
 and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.0](opera-rtc-s1-browse/compare/v0.0.0...v0.1.0) -- unreleased
+## [0.1.0]
 
+### Added
+* RTC granule download functionality
+* Browse image creation functionality
+
+### Removed
+* Unused functionality created during HyP3-Cookiecutter setup